+++ allennlp/tests/nn/util_test.py
@@ -463,7 +463,7 @@ class TestNnUtil(AllenNlpTestCase):
             }
         }
         assert_almost_equal(
-            util.get_text_field_mask(text_field_tensors).numpy(), [[1, 1, 1, 0, 0], [1, 1, 0, 0, 0]]
+            util.get_text_field_mask(text_field_tensors).long().numpy(), [[1, 1, 1, 0, 0], [1, 1, 0, 0, 0]]
         )
 
     def test_get_text_field_mask_returns_a_correct_mask_character_only_input(self):
@@ -478,7 +478,7 @@ class TestNnUtil(AllenNlpTestCase):
             }
         }
         assert_almost_equal(
-            util.get_text_field_mask(text_field_tensors).numpy(), [[1, 1, 1, 0], [1, 1, 0, 0]]
+            util.get_text_field_mask(text_field_tensors).long().numpy(), [[1, 1, 1, 0], [1, 1, 0, 0]]
         )
 
     def test_get_text_field_mask_returns_a_correct_mask_list_field(self):
@@ -492,7 +492,7 @@ class TestNnUtil(AllenNlpTestCase):
                 )
             }
         }
-        actual_mask = util.get_text_field_mask(text_field_tensors, num_wrapping_dims=1).numpy()
+        actual_mask = util.get_text_field_mask(text_field_tensors, num_wrapping_dims=1).long().numpy()
         expected_mask = (text_field_tensors["indexer_name"]["list_tokens"].numpy() > 0).astype(
             "int32"
         )
@@ -505,7 +505,7 @@ class TestNnUtil(AllenNlpTestCase):
                 "mask": torch.LongTensor([[0, 0, 1]]),
             }
         }
-        assert_almost_equal(util.get_text_field_mask(text_field_tensors).numpy(), [[0, 0, 1]])
+        assert_almost_equal(util.get_text_field_mask(text_field_tensors).long().numpy(), [[0, 0, 1]])
 
     def test_weighted_sum_works_on_simple_input(self):
         batch_size = 1
